anet-services/requirements.txt
anet-sdk
fastapi
uvicorn
anthropic
